apps/services.codapdf.com/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.39
botocore==1.35.39
cachelib==0.9.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
distlib==0.3.9
filelock==3.16.1
Flask==3.0.3
Flask-Caching==2.3.0
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==20.1.0
htmlmin==0.1.12
identify==2.6.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
multidict==6.1.0
mypy==1.13.0
mypy-extensions==1.0.0
nodeenv==1.9.1
platformdirs==4.3.6
playwright==1.47.0
propcache==0.2.0
pyee==12.0.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rcssmin==1.1.3
requests==2.32.3
rjsmin==1.2.3
s3transfer==0.10.3
setuptools==75.2.0
six==1.16.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==1.26.15
virtualenv==20.27.0
Werkzeug==3.0.4
yarl==1.15.5
